requirements.txt
aiohttp==3.7.4.post0
anyio==3.5.0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
deprecation==2.1.0
discord.py==1.7.3
gotrue==0.4.0
h11==0.12.0
httpcore==0.14.5
httpx==0.21.3
idna==3.3
multidict==5.2.0
packaging==21.3
postgrest-py==0.8.0
pydantic==1.9.0
pyparsing==3.0.6
python-dateutil==2.8.2
realtime==0.0.4
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
supabase==0.3.0
typing-extensions==4.0.1
urllib3==1.26.8
websockets==9.1
yarl==1.7.2
